requirements.txt
requests

spotipy

python-dotenv

IPython

pandas

plotly

pytest

flask

gunicorn